bitcoin-sopr-dashboard/requirements.txt
yfinance>=0.2.28
google-cloud-bigquery>=3.11.0
db-dtypes>=1.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
plotly>=5.17.0
